Code Challenge 08.18
Given 2 strings, a and b, return the number of the positions where they contain the same 2-character substrings. So "xxcaazz" and "xxbaaz" yields 3, since the "xx", "aa", and "az" substrings appear in the same indices in both strings. You can assume that the string arguments will be the same length.