Code Challenge 08.19
The squirrels in Palo Alto spend most of the day playing. In particular, they play if the temperature is between 60 and 90 (inclusive). Unless it is summer, then they'll play when it is up to 100 instead of 90. Given an int Fahrenheit temperature and a boolean is_summer, return True if the squirrels play and False otherwise.

squirrel_play(70, False) → True

squirrel_play(95, False) → False

squirrel_play(95, True) → True